+++ spec/lite/ruby/configuration_spec.rb
@@ -6,8 +6,8 @@ RSpec.describe Lite::Ruby::Configuration do
   after do
     Lite::Ruby.configure do |config|
       config.monkey_patches = %w[
-        array date enumerable hash integer kernel numeric object open_struct range string struct
-        time
+        array boolean date enumerable hash integer kernel numeric object open_struct range string
+        struct time
       ]
     end
   end